+++ pkg/storegateway/pool.go
@@ -6,10 +6,17 @@
 package storegateway
 
 import (
+	"sync"
+
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
 
+	"github.com/grafana/mimir/pkg/storegateway/storepb"
 	"github.com/grafana/mimir/pkg/util/pool"
+)
+
+const (
+	chunkSliceSize = 16_384 // An arbitrary size limit, susceptible to change.
 )
 
 type chunkBytesPool struct {
@@ -53,4 +60,52 @@ func (p *chunkBytesPool) Get(sz int) (*[]byte, error) {
 
 func (p *chunkBytesPool) Put(b *[]byte) {
 	p.pool.Put(b)
+}
+
+// chunkSlicePool is a memory pool of chunk slices.
+type chunkSlicePool struct {
+	pool sync.Pool
+}
+
+// newChunksSlicePool creates a new chunks pool.
+func newChunksSlicePool() *chunkSlicePool {
+	return &chunkSlicePool{
+		pool: sync.Pool{
+			New: func() interface{} {
+				return &chunkSlice{
+					chunks: make([]storepb.Chunk, chunkSliceSize),
+				}
+			},
+		},
+	}
+}
+
+// get returns a chunk slice from the pool.
+func (p *chunkSlicePool) get() *chunkSlice {
+	return p.pool.Get().(*chunkSlice)
+}
+
+// put returns a chunk slice to the pool.
+func (p *chunkSlicePool) put(slice *chunkSlice) {
+	slice.i = 0
+	p.pool.Put(slice)
+}
+
+type chunkSlice struct {
+	chunks []storepb.Chunk
+	i      int
+}
+
+func (s *chunkSlice) next() *storepb.Chunk {
+	if s.isExhausted() {
+		return nil
+	}
+	chk := &s.chunks[s.i]
+	chk.Data = nil
+	s.i++
+	return chk
+}
+
+func (s *chunkSlice) isExhausted() bool {
+	return s.i >= len(s.chunks)
 }